PVGrammar.Prob.Simple

Description

This module contains a simple (and musically rather naive) probabilistic model of protovoice derivations. This model can be used to sample a derivation, evaluate a derivations probability, or infer posterior distributions of the model parmeters from given derivations (i.e., "learn" the model's probabilities).

This model is a locally conjugate model: It samples a derivation using a sequence of random decisions with certain probabilities. These probabilities are generally unknown, so they are themselves modeled as random variables with prior distributions. The full model \(p(d, \theta)\) thus splits into \[p(D, \theta) = p(d \mid \theta) \cdot p(\theta),\] the prior over the probability variables \[p(\theta) = \prod_i p(\theta_i),\] and the likelihood of the derivation(s) given these probabilities \[p(D \mid \theta) = \prod_{d \in D} p(d \mid \theta) = \prod_{d \in D} \prod_i p(d_i \mid \theta, d_0, \ldots, d_{i-1}).\] Given all prior decisions, the likelihood of a decision \(d_i\) based on some parameter \(\theta_a\) \[p(d_i \mid \theta, d_{<i})\] is conjugate with the prior of that parameter \(p(\theta_a)\), which means that the posterior of the parameters given one (or several) derivation(s) \(p(\theta \mid D)\) can be computed analytically.

The parameters \(\theta\) and their prior distributions are represented by the higher-kinded type PVParams. Different instantiations of this type (using Hyper or Probs) results in concrete record types that represent prior or posterior distributions or concrete values (probabilities) for the parameters. PVParams also supports jeffreysPrior and uniformPrior as default priors, as well as sampleProbs for sampling from a prior (see Inferenc.Conjugate).

The likelihood \(p(d \mid \theta)\) of a derivation is represented by sampleDerivation. It can be executed under different "modes" (probability monads) for sampling, inference, or tracing (see Inference.Conjugate). The decisions during the derivation are represented by a Trace (here Trace PVParams). In order to learn from a given derivation, the corresponding trace can be obtained using observeDerivation. A combination of getting a trace and learning from it is provided by trainSinglePiece.

Model Parameters

A higher-kinded type that represents the global parameters (probabilities) of the model. Use it as 'Hyper PVParams' to represent hyperparameters (priors and posteriors) or as 'Probs PVParams' to represent actual probabilites. Each record field corresponds to one parameter that influences a specific type of decision in the generation process.

data PVParams (f :: Type -> Type)

The combined parameters for inner and outer operations.

Constructors

PVParams

Fields

_pOuter :: PVParamsOuter f

_pInner :: PVParamsInner f

data PVParamsOuter (f :: Type -> Type)

Parameters for decisions about outer operations (split, spread, freeze).

Constructors

PVParamsOuter

Fields

_pSingleFreeze :: f Beta

_pDoubleLeft :: f Beta

_pDoubleLeftFreeze :: f Beta

_pDoubleRightSplit :: f Beta

data PVParamsInner (f :: Type -> Type)

Parameters for decisions about inner operations (elaboration and distribution within splits and spreads).

Constructors

PVParamsInner

Fields

_pElaborateRegular :: f Beta

_pElaborateL :: f Beta

_pElaborateR :: f Beta

_pRootFifths :: f Beta

_pKeepL :: f Beta

_pKeepR :: f Beta

_pRepeatOverNeighbor :: f Beta

_pNBChromatic :: f Beta

_pNBAlt :: f Beta

_pRepeatLeftOverRight :: f Beta

_pRepeatAlter :: f Beta

_pRepeatAlterUp :: f Beta

_pRepeatAlterSemis :: f Beta

_pConnect :: f Beta

_pConnectChromaticLeftOverRight :: f Beta

_pPassUp :: f Beta

_pPassLeftOverRight :: f Beta

_pNewPassingLeft :: f Beta

_pNewPassingRight :: f Beta

_pNewPassingMid :: f Beta

_pNoteSpreadDirection :: f (Dirichlet 3)

_pNotesOnOtherSide :: f Beta

_pSpreadRepetitionEdge :: f Beta

savePVHyper :: FilePath -> Hyper PVParams -> IO ()

loadPVHyper :: FilePath -> IO (Either String (Hyper PVParams))

Likelihood Model

sampleDerivation represents a probabilistic program that samples a derivation. that can be interpreted in various modes for

sampling (sampleTrace, sampleResult),

inference (evalTraceLogP, getPosterior),

tracing (showTrace, traceTrace).

observeDerivation takes and existing derivation and returns the corresponding trace.

sampleDerivation

Arguments

:: (SampleCtx m Bernoulli, SampleCtx m Geometric1, SampleCtx m Geometric0, SampleCtx m MagicalOctaves, SampleCtx m MagicalID, SampleCtx m (Categorical 3), RandomInterpreter m PVParams)

=> Path (Edges SPitch) (Notes SPitch)
root path

-> m (Either String [PVLeftmost SPitch])
a probabilistic program

A probabilistic program that samples a derivation starting from a given root path. Can be interpreted by the interpreter functions in Inference.Conjugate.

sampleDerivation' :: (SampleCtx m Bernoulli, SampleCtx m Geometric1, SampleCtx m Geometric0, SampleCtx m MagicalOctaves, SampleCtx m MagicalID, SampleCtx m (Categorical 3), RandomInterpreter m PVParams) => m (Either String [PVLeftmost SPitch])

A shorthand for sampleDerivation starting from ⋊——⋉.

observeDerivation :: [PVLeftmost SPitch] -> Path (Edges SPitch) (Notes SPitch) -> Either String (Trace PVParams)

Walk through a derivation (starting at a given root path) and return the corresponding Trace (if possible). The trace can be used together with sampleDerivation for inference (getPosterior) or for showing the trace (printTrace).

observeDerivation' :: [PVLeftmost SPitch] -> Either String (Trace PVParams)

A shorthand for observeDerivation starting from ⋊——⋉.

Utilities

roundtrip :: FilePath -> IO (Either String ())

A helper function that tests whether observeDerivation' followed by sampleDerivation' restores the original derivation. Useful for testing the compatibility of the two functions.

trainSinglePiece :: FilePath -> IO (Maybe (PVParams HyperRep))

Helper function: Load a single derivation and infer the corresponding posterior for a uniform prior.

Likelihood model for parsing

We need these specialized functions because of a dependency across steps: During a double step (elaborating two transitions), the process must decide whether to elaborate the left or right transition. To normalize the derivation order, we can't elaborate the left transition after the right one. That means that the model *sometimes* has to make the decision to go right, and sometimes not (i.e., after a right split). During parsing, we don't know whether this decision had to be made or not, since we don't know the previous derivation step yet. Therefore, we don't include the decision in the current step, but at the end of the previous one (in generation order), where the context is known. As a consequence, the result of this decision (if made) needs to be passed to the functions scoring the previous step. When parsing, make sure to maintain this information.

sampleSingleStepParsing :: (SampleCtx m Bernoulli, SampleCtx m Geometric1, SampleCtx m Geometric0, SampleCtx m MagicalOctaves, SampleCtx m MagicalID, RandomInterpreter m PVParams) => ContextSingle SPitch -> m ()

Sample a single step in a bottom-up context. Only used for evaluating the probability of a step, therefore returns ().

observeSingleStepParsing

Arguments

:: ContextSingle SPitch
the parent path

-> Maybe Bool
If the following (generative) step is a double op, this is the result of the "continueLeft" decision, i.e., 'Just True' for split-left and freeze-left, and 'Just False' for spread and split-right. If the following step is a single op, this is Nothing (as no decision is made).

-> LeftmostSingle (Split SPitch) (Freeze SPitch)
the performed operation

-> Either String (Trace PVParams)

Observerse a single step in a bottom-up context. Since double operations don't know whether they have to make a "continueLeft" decision when going bottom-up, this decision is moved to the previous step, where the context is know. Therefore, if the following step would have to make this decision, it is added here.

evalSingleStep :: Probs PVParams -> ContextSingle SPitch -> LeftmostSingle (Split SPitch) (Freeze SPitch) -> Maybe Bool -> Either String (Maybe ((), Double))

sampleDoubleStepParsing :: (SampleCtx m Bernoulli, SampleCtx m Geometric1, SampleCtx m Geometric0, SampleCtx m MagicalOctaves, SampleCtx m MagicalID, SampleCtx m (Categorical 3), RandomInterpreter m PVParams) => ContextDouble SPitch -> LeftmostDouble (Split SPitch) (Freeze SPitch) (Spread SPitch) -> m ()

Sample a double step in a bottom-up context. Only used for evaluating the probability of a step, therefore takes the "resulting" op and returns ().

observeDoubleStepParsing

Arguments

:: ContextDouble SPitch
the parent path

-> Maybe Bool
If the following (generative) step is a double op, this is the result of the "continueLeft" decision, i.e., 'Just True' for split-left and freeze-left and 'Just False' for spread and split-right. If the following step is a single op, this is Nothing, as not decision is made.

-> LeftmostDouble (Split SPitch) (Freeze SPitch) (Spread SPitch)
the performed operation

-> Either String (Trace PVParams)

Observerse a double step without knowing if it happened after a right split (e.g., when parsing). The extra decision that is necessary if it doesn't follow a right split is "moved" to the previous step. Therefore, this step is rated as if it follows a right split (not making the decision). In addition, if the following step would have to make the extra decision, it is added here.

evalDoubleStep :: Probs PVParams -> ContextDouble SPitch -> LeftmostDouble (Split SPitch) (Freeze SPitch) (Spread SPitch) -> Maybe Bool -> Either String (Maybe ((), Double))

Orphan instances

FromJSON (HyperRep Beta)

Instance details

Methods

parseJSON :: Value -> Parser (HyperRep Beta)

parseJSONList :: Value -> Parser [HyperRep Beta]

omittedField :: Maybe (HyperRep Beta)

FromJSON (HyperRep (Dirichlet 3))

Instance details

Methods

parseJSON :: Value -> Parser (HyperRep (Dirichlet 3))

parseJSONList :: Value -> Parser [HyperRep (Dirichlet 3)]

omittedField :: Maybe (HyperRep (Dirichlet 3))

ToJSON (HyperRep Beta)

Instance details

Methods

toJSON :: HyperRep Beta -> Value

toEncoding :: HyperRep Beta -> Encoding

toJSONList :: [HyperRep Beta] -> Value

toEncodingList :: [HyperRep Beta] -> Encoding

omitField :: HyperRep Beta -> Bool

ToJSON (HyperRep (Dirichlet 3))

Instance details

Methods

toJSON :: HyperRep (Dirichlet 3) -> Value

toEncoding :: HyperRep (Dirichlet 3) -> Encoding

toJSONList :: [HyperRep (Dirichlet 3)] -> Value

toEncodingList :: [HyperRep (Dirichlet 3)] -> Encoding

omitField :: HyperRep (Dirichlet 3) -> Bool

Generic (HyperRep Beta)

Instance details

Methods

from :: HyperRep Beta -> Rep (HyperRep Beta) x

to :: Rep (HyperRep Beta) x -> HyperRep Beta

Generic (HyperRep (Dirichlet 3))

Instance details

Associated Types

type Rep (HyperRep (Dirichlet 3))

Instance details

Defined in PVGrammar.Prob.Simple

type Rep (HyperRep (Dirichlet 3)) = D1 ('MetaData "HyperRep" "Inference.Conjugate" "conjugate-programs-0.1.0.0-3yRv9wXBSRgCze1zWxB2WM" 'True) (C1 ('MetaCons "HyperRep" 'PrefixI 'True) (S1 ('MetaSel ('Just "runHyper") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Hyper (AsPrior (Dirichlet 3))))))

type Rep (HyperRep (Dirichlet 3))

Instance details

Defined in PVGrammar.Prob.Simple

type Rep (HyperRep (Dirichlet 3)) = D1 ('MetaData "HyperRep" "Inference.Conjugate" "conjugate-programs-0.1.0.0-3yRv9wXBSRgCze1zWxB2WM" 'True) (C1 ('MetaCons "HyperRep" 'PrefixI 'True) (S1 ('MetaSel ('Just "runHyper") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Hyper (AsPrior (Dirichlet 3))))))

Methods

from :: HyperRep (Dirichlet 3) -> Rep (HyperRep (Dirichlet 3)) x

to :: Rep (HyperRep (Dirichlet 3)) x -> HyperRep (Dirichlet 3)